+++ src/utils/index.ts
@@ -1,3 +1,4 @@
+export * from './computeJwkThumbprint';
 export * from './extractRawPrivateKeyFromPkcs8';
 export * from './getErrorMessage';
 export * from './resolveAlgorithmName';